zaps/admin_users/core/fastapi/middlewares/authentication.py
from typing import Optional, Tuple

import jwt
from starlette.authentication import AuthenticationBackend
from starlette.middleware.authentication import (
    AuthenticationMiddleware as BaseAuthenticationMiddleware,
)
from starlette.requests import HTTPConnection

from core.config import config
from ..schemas import CurrentUser


class AuthenticationMiddleware(BaseAuthenticationMiddleware):
    pass
