zaps/admin_users/requirements.txt
anyio==3.6.2
asyncpg==0.27.0
backports.zoneinfo==0.2.1
certifi==2022.12.7
charset-normalizer==3.0.1
click==8.1.3
databases==0.7.0
dnspython==2.3.0
email-validator==1.3.1
fastapi==0.91.0
greenlet==2.0.2
h11==0.14.0
idna==3.4
psycopg==3.1.8
psycopg2-binary==2.9.5
pyasn1==0.4.8
pydantic==1.10.4
python-dotenv==0.21.1
python-multipart==0.0.5
requests==2.28.2
rsa==4.9
six==1.16.0
sniffio==1.3.0
SQLAlchemy==2.0.3
starlette==0.24.0
typing-extensions==4.4.0
urllib3==1.26.14
uvicorn==0.20.0
